# services/gateway/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.InspectionResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      id:
        type: string
      inspector_id:
        type: string
      notes:
        type: string
      permit_id:
        type: string
      result:
        type: string
      scheduled_at:
        type: string
    type: object
...
  handlers.UserResponse:
    properties:
      created_at:
        type: string
      department_id:
        type: string
      email:
        type: string
      id:
        type: string
      oidc_sub:
        type: string
      role:
        type: string
    type: object
...
  handlers.createUserInput:
    properties:
      department_id:
        type: string
      email:
        type: string
      oidc_sub:
        type: string
      role:
        type: string
    type: object
...
  handlers.scheduleInspectionRequest:
    properties:
      inspector_id:
        type: string
      scheduled_at:
        type: string
    type: object
  handlers.updateInspectionRequest:
    properties:
      completed_at:
        type: string
      notes:
        type: string
      result:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Fieldstone Contributors
    url: https://github.com/fieldstone/fieldstone
  description: |-
    A self-hostable platform for city governments to run common civic services.

    Staff endpoints require a Bearer JWT from the configured OIDC provider.
    Citizen endpoints (marked "public") do not require authentication.

    **Development**: set `DEV_DISABLE_AUTH=true` in .env to bypass JWT
    validation locally. Remove before production deployment.
  license:
    name: MIT
    url: https://github.com/fieldstone/fieldstone/blob/main/LICENSE
  title: Fieldstone Civic Services API
  version: 0.1.0
paths:
...
  /v1/permits/{id}/inspections:
    post:
      consumes:
      - application/json
      parameters:
      - description: Permit UUID
        in: path
        name: id
        required: true
        type: string
      - description: Inspection details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.scheduleInspectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.InspectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Schedule a permit inspection
      tags:
      - permits
  /v1/permits/{id}/inspections/{iid}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Permit UUID
        in: path
        name: id
        required: true
        type: string
      - description: Inspection UUID
        in: path
        name: iid
        required: true
        type: string
      - description: Inspection outcome
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateInspectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.InspectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an inspection result
      tags:
      - permits
...
  /v1/users:
    get:
      parameters:
      - description: Filter by department UUID
        in: query
        name: department_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handlers.UserResponse'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List staff users
      tags:
      - identity
    post:
      consumes:
      - application/json
      description: |-
        Links a city employee's OIDC identity to a department and role.
        Publishes identity.user.provisioned.
      parameters:
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: oidc_sub already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Provision a staff user
      tags:
      - identity
  /v1/users/me:
    get:
      description: |-
        Returns the staff user record matching the JWT subject claim.
        In development (DEV_DISABLE_AUTH=true) returns a synthetic dev user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "404":
          description: User not yet provisioned
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - identity
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token: `Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"